fix: remove format and/or codec options
IPFS is all `dag-pb` yet we allow passing `dag-cbor` as the format of
IPLD node to use.  This is likely to break anything that uses it as
tools expect IPFS stuff to be `dag-pb`, possibly with `raw` leaves.

Closes #67

BREAKING CHANGE:

`--format` and/or `--codec` option has been removed from the CLI, the
HTTP API and the core API.